Add school year data, clean structure
