+++ src/i18n/translations/en.js
@@ -15,6 +15,7 @@ export default {
   'unit.contact.info': 'Contact information',
   'unit.services': 'Services',
   'unit.e.services': 'Electronic services', // TODO: verify
+  'unit.school.year': 'School year',
   'unit.opens.new.tab': '(new tab)',
   'unit.call.number': '(call)',
